+++ Assets/DeltaDNA/Helpers/Utils.cs
@@ -4,9 +4,8 @@ using System.Collections.Generic;
 using System.IO;
 
 #if NETFX_CORE
-using UnityEngine.Windows;
-using Windows.Storage;
 using System.Threading.Tasks;
+using Windows.Storage;
 using Windows.Storage.Streams;
 #endif
 
@@ -42,16 +41,40 @@ namespace DeltaDNA
 			#endif
 		}
 
+		public static bool FileExists(string path)
+		{
+			#if UNITY_WINRT
+			return UnityEngine.Windows.File.Exists(path);
+			#elif UNITY_WEBPLAYER || UNITY_WEBGL
+			return false; // since we won't actually make a file
+			#else
+			return System.IO.File.Exists(path);
+			#endif
+		}
+
+		public static bool DirectoryExists(string path)
+		{
+			#if UNITY_WINRT
+			return UnityEngine.Windows.Directory.Exists(path);
+			#elif UNITY_WEBPLAYER || UNITY_WEBGL
+			return false; // since we won't actually make a file
+			#else
+			return System.IO.Directory.Exists(path);
+			#endif
+		}
+
         public static void CreateDirectory(string path)
         {
-            #if NETFX_CORE
+			#if UNITY_WINRT
             // Unity's WP8.1 version from Windows.Storage doesn't do it recursively
             path = path.Replace('/', '\\');
             string parent = path.Substring(0, path.LastIndexOf('\\'));
-            if (!Directory.Exists(parent)) {
+            if (!UnityEngine.Windows.Directory.Exists(parent)) {
                 CreateDirectory(parent);
             }
             UnityEngine.Windows.Directory.CreateDirectory(path);
+			#elif UNITY_WEBPLAYER || UNITY_WEBGL
+			return;
             #else
             System.IO.Directory.CreateDirectory(path);
             #endif
@@ -70,7 +93,7 @@ namespace DeltaDNA
 
             throw thread.Exception;
 
-            #elif UNITY_WEBPLAYER
+            #elif UNITY_WEBPLAYER || UNITY_WEBGL
             Logger.LogDebug("Creating memory based stream");
             return new MemoryStream();
             #else
@@ -91,7 +114,7 @@ namespace DeltaDNA
                 return thread.Result;
 
             throw thread.Exception;
-            #elif UNITY_WEBPLAYER
+            #elif UNITY_WEBPLAYER || UNITY_WEBGL
             Logger.LogDebug("Opening memory based stream");
             return new MemoryStream();
             #else